+++ src/main/java/com/github/romankh3/droolsspringboot/DroolsConfig.java
@@ -1,0 +1,56 @@
+package com.github.romankh3.droolsspringboot;
+
+import java.io.IOException;
+import org.kie.api.KieServices;
+import org.kie.api.builder.KieBuilder;
+import org.kie.api.builder.KieFileSystem;
+import org.kie.api.builder.KieModule;
+import org.kie.api.builder.KieRepository;
+import org.kie.api.builder.ReleaseId;
+import org.kie.api.runtime.KieContainer;
+import org.kie.api.runtime.KieSession;
+import org.kie.internal.io.ResourceFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class DroolsConfig {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(DroolsConfig.class);
+
+    private KieServices kieServices = KieServices.Factory.get();
+
+    @Bean
+    public KieSession getKieSession() throws IOException {
+        LOGGER.info("Session created...");
+        return getKieContainer().newKieSession();
+    }
+
+    @Bean
+    public KieContainer getKieContainer() throws IOException {
+        LOGGER.info("Container created...");
+        getKieRepository();
+        KieBuilder kb = kieServices.newKieBuilder(getKieFileSystem());
+        kb.buildAll();
+        KieModule kieModule = kb.getKieModule();
+        return kieServices.newKieContainer(kieModule.getReleaseId());
+    }
+
+    private void getKieRepository() {
+        final KieRepository kieRepository = kieServices.getRepository();
+        kieRepository.addKieModule(new KieModule() {
+            @Override
+            public ReleaseId getReleaseId() {
+                return kieRepository.getDefaultReleaseId();
+            }
+        });
+    }
+
+    private KieFileSystem getKieFileSystem() throws IOException {
+        KieFileSystem kieFileSystem = kieServices.newKieFileSystem();
+        kieFileSystem.write(ResourceFactory.newClassPathResource("order.drl"));
+        return kieFileSystem;
+    }
+}