+++ src/test/java/com/github/romankh3/droolsspringboot/ApplicationIT.java
@@ -1,0 +1,15 @@
+package com.github.romankh3.droolsspringboot;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ApplicationIT {
+
+    @Test
+    public void shouldLoadApplicationContext() {
+    }
+}